rename pets files
