+++ subgraphs/smolverse/src/mappings/smol-bodies-pets.ts
